+++ programs/MushroomClassifier.php
@@ -9,6 +9,7 @@ use Rubix\ML\Reports\AggregateReport;
 use Rubix\ML\Reports\ConfusionMatrix;
 use Rubix\ML\Reports\PredictionSpeed;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
+use Rubix\ML\NeuralNet\Initializers\He;
 use Rubix\ML\Transformers\OneHotEncoder;
 use Rubix\ML\Reports\MulticlassBreakdown;
 use Rubix\ML\Transformers\LambdaFunction;
@@ -47,13 +48,13 @@ $labels = iterator_to_array($reader->fetchColumn('class'));
 $dataset = new Labeled($samples, $labels);
 
 $estimator = new Pipeline(new MultiLayerPerceptron([
-    new Dense(30),
+    new Dense(30, new He()),
     new Activation(new SELU()),
     new AlphaDropout(0.1),
-    new Dense(20),
+    new Dense(20, new He()),
     new Activation(new SELU()),
     new AlphaDropout(0.1),
-    new Dense(10),
+    new Dense(10, new He()),
     new Activation(new SELU()),
 ], 100, new Adam(0.001), 1e-4, new CrossEntropy(), 1e-3, new MCC(), 0.1, 3, 100), [
     new OneHotEncoder(),