+++ src/NeuralNet/Layers/Continuous.php
@@ -178,7 +178,7 @@ class Continuous implements Output
         $this->weights->update($optimizer->step($this->weights, $dW));
         $this->biases->update($optimizer->step($this->biases, $dB));
 
-        $cost = $delta->sum()->sum();
+        $cost = $delta->sum()->mean();
 
         unset($this->input, $this->z);