+++ src/NeuralNet/Layers/Multiclass.php
@@ -228,7 +228,7 @@ class Multiclass implements Output
         $this->weights->update($optimizer->step($this->weights, $dW));
         $this->biases->update($optimizer->step($this->biases, $dB));
 
-        $cost = $delta->sum()->sum();
+        $cost = $delta->sum()->mean();
 
         unset($this->input, $this->z, $this->computed);